Moves test batch-notification config files to correct folder
