+++ test/modules/batch-notifications/config/returns/lib/get-recipients.js
@@ -7,18 +7,18 @@ const {
 } = exports.lab = require('@hapi/lab').script();
 const sandbox = require('sinon').createSandbox();
 
-const { logger } = require('../../../../../../../src/logger');
-const Contact = require('../../../../../../../src/lib/models/contact');
+const { logger } = require('../../../../../../src/logger');
+const Contact = require('../../../../../../src/lib/models/contact');
 const notificationContacts =
-  require('../../../../../../../src/modules/batch-notifications/config/returns/lib/return-notification-contacts');
+  require('../../../../../../src/modules/batch-notifications/config/returns/lib/return-notification-contacts');
 const notificationRecipients =
-  require('../../../../../../../src/modules/batch-notifications/config/returns/lib/return-notification-recipients');
+  require('../../../../../../src/modules/batch-notifications/config/returns/lib/return-notification-recipients');
 const { getRecipients } =
-  require('../../../../../../../src/modules/batch-notifications/config/returns/lib/get-recipients');
+  require('../../../../../../src/modules/batch-notifications/config/returns/lib/get-recipients');
 const scheduledNotifications =
-  require('../../../../../../../src/controllers/notifications');
+  require('../../../../../../src/controllers/notifications');
 const eventHelpers =
-  require('../../../../../../../src/modules/batch-notifications/lib/event-helpers');
+  require('../../../../../../src/modules/batch-notifications/lib/event-helpers');
 
 experiment('getRecipients', () => {
   const jobData = {
